+++ library/common/src/main/java/com/google/android/exoplayer2/Format.java
@@ -49,12 +49,50 @@ public final class Format implements Parcelable {
   @Nullable public final String id;
   /** The human readable label, or null if unknown or not applicable. */
   @Nullable public final String label;
+  /** The language as an IETF BCP 47 conformant tag, or null if unknown or not applicable. */
+  @Nullable public final String language;
   /** Track selection flags. */
   @C.SelectionFlags public final int selectionFlags;
   /** Track role flags. */
   @C.RoleFlags public final int roleFlags;
   /**
-   * The average bandwidth in bits per second, or {@link #NO_VALUE} if unknown or not applicable.
+   * The average bitrate in bits per second, or {@link #NO_VALUE} if unknown or not applicable. This
+   * field may be populated from the following sources, depending on media type and the type of the
+   * track within the media:
+   *
+   * <ul>
+   *   <li>DASH: Always {@link Format#NO_VALUE}.
+   *   <li>HLS: The {@code AVERAGE-BANDWIDTH} attribute defined on the corresponding {@code
+   *       EXT-X-STREAM-INF} tag in the master playlist, or {@link Format#NO_VALUE} if not present.
+   *   <li>SmoothStreaming: The {@code Bitrate} attribute defined on the corresponding {@code
+   *       TrackElement} in the manifest, or {@link Format#NO_VALUE} if not present.
+   *   <li>Progressive container formats: Often {@link Format#NO_VALUE}, but may be populated with
+   *       the average bitrate if defined by the container.
+   * </ul>
+   */
+  // TODO: Make public.
+  private final int averageBitrate;
+  /**
+   * The peak bitrate in bits per second, or {@link #NO_VALUE} if unknown or not applicable. This
+   * field may be populated from the following sources, depending on media type and the type of the
+   * track within the media:
+   *
+   * <ul>
+   *   <li>DASH: The {@code @bandwidth} attribute of the corresponding {@code Representation}
+   *       element in the manifest.
+   *   <li>HLS: The {@code BANDWIDTH} attribute defined on the corresponding {@code
+   *       EXT-X-STREAM-INF} tag.
+   *   <li>SmoothStreaming: Always {@link Format#NO_VALUE}.
+   *   <li>Progressive container formats: Often {@link Format#NO_VALUE}, but may be populated with
+   *       the peak bitrate if defined by the container.
+   * </ul>
+   */
+  // TODO: Make public.
+  private final int peakBitrate;
+  /**
+   * The bitrate in bits per second. This is the peak bitrate if known, or else the average bitrate
+   * if known, or else {@link Format#NO_VALUE}. Equivalent to: {@code peakBitrate != NO_VALUE ?
+   * peakBitrate : averageBitrate}.
    */
   public final int bitrate;
   /** Codecs of the format as described in RFC 6381, or null if unknown or not applicable. */
@@ -149,13 +187,9 @@ public final class Format implements Parcelable {
    */
   public final int encoderPadding;
 
-  // Audio and text specific.
+  // Text specific.
 
-  /** The language as an IETF BCP 47 conformant tag, or null if unknown or not applicable. */
-  @Nullable public final String language;
-  /**
-   * The Accessibility channel, or {@link #NO_VALUE} if not known or applicable.
-   */
+  /** The Accessibility channel, or {@link #NO_VALUE} if not known or applicable. */
   public final int accessibilityChannel;
 
   // Provided by source.
@@ -189,9 +223,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        /* language= */ null,
         selectionFlags,
         roleFlags,
-        bitrate,
+        /* averageBitrate= */ bitrate,
+        /* peakBitrate= */ bitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -213,7 +249,6 @@ public final class Format implements Parcelable {
         /* pcmEncoding= */ NO_VALUE,
         /* encoderDelay= */ NO_VALUE,
         /* encoderPadding= */ NO_VALUE,
-        /* language= */ null,
         /* accessibilityChannel= */ NO_VALUE,
         /* exoMediaCryptoType= */ null);
   }
@@ -294,9 +329,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         /* label= */ null,
+        /* language= */ null,
         /* selectionFlags= */ 0,
         /* roleFlags= */ 0,
-        bitrate,
+        /* averageBitrate= */ bitrate,
+        /* peakBitrate= */ bitrate,
         codecs,
         /* metadata= */ null,
         /* containerMimeType= */ null,
@@ -318,7 +355,6 @@ public final class Format implements Parcelable {
         /* pcmEncoding= */ NO_VALUE,
         /* encoderDelay= */ NO_VALUE,
         /* encoderPadding= */ NO_VALUE,
-        /* language= */ null,
         /* accessibilityChannel= */ NO_VALUE,
         /* exoMediaCryptoType= */ null);
   }
@@ -342,9 +378,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        /* averageBitrate= */ bitrate,
+        /* peakBitrate= */ bitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -366,7 +404,6 @@ public final class Format implements Parcelable {
         /* pcmEncoding= */ NO_VALUE,
         /* encoderDelay= */ NO_VALUE,
         /* encoderPadding= */ NO_VALUE,
-        language,
         /* accessibilityChannel= */ NO_VALUE,
         /* exoMediaCryptoType= */ null);
   }
@@ -448,9 +485,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         /* label= */ null,
+        language,
         selectionFlags,
         /* roleFlags= */ 0,
-        bitrate,
+        /* averageBitrate= */ bitrate,
+        /* peakBitrate= */ bitrate,
         codecs,
         metadata,
         /* containerMimeType= */ null,
@@ -472,7 +511,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         /* accessibilityChannel= */ NO_VALUE,
         /* exoMediaCryptoType= */ null);
   }
@@ -516,9 +554,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        /* averageBitrate= */ bitrate,
+        /* peakBitrate= */ bitrate,
         codecs,
         /* metadata= */ null,
         containerMimeType,
@@ -540,7 +580,6 @@ public final class Format implements Parcelable {
         /* pcmEncoding= */ NO_VALUE,
         /* encoderDelay= */ NO_VALUE,
         /* encoderPadding= */ NO_VALUE,
-        language,
         accessibilityChannel,
         /* exoMediaCryptoType= */ null);
   }
@@ -571,9 +610,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         /* label= */ null,
+        language,
         selectionFlags,
         /* roleFlags= */ 0,
-        /* bitrate= */ NO_VALUE,
+        /* averageBitrate= */ NO_VALUE,
+        /* peakBitrate= */ NO_VALUE,
         /* codecs= */ null,
         /* metadata= */ null,
         /* containerMimeType= */ null,
@@ -595,7 +636,6 @@ public final class Format implements Parcelable {
         /* pcmEncoding= */ NO_VALUE,
         /* encoderDelay= */ NO_VALUE,
         /* encoderPadding= */ NO_VALUE,
-        language,
         accessibilityChannel,
         /* exoMediaCryptoType= */ null);
   }
@@ -611,9 +651,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         /* label= */ null,
+        language,
         selectionFlags,
         /* roleFlags= */ 0,
-        /* bitrate= */ NO_VALUE,
+        /* averageBitrate= */ NO_VALUE,
+        /* peakBitrate= */ NO_VALUE,
         /* codecs= */ null,
         /* metadata=*/ null,
         /* containerMimeType= */ null,
@@ -635,7 +677,6 @@ public final class Format implements Parcelable {
         /* pcmEncoding= */ NO_VALUE,
         /* encoderDelay= */ NO_VALUE,
         /* encoderPadding= */ NO_VALUE,
-        language,
         /* accessibilityChannel= */ NO_VALUE,
         /* exoMediaCryptoType= */ null);
   }
@@ -655,9 +696,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        /* averageBitrate= */ bitrate,
+        /* peakBitrate= */ bitrate,
         codecs,
         /* metadata= */ null,
         containerMimeType,
@@ -679,7 +722,6 @@ public final class Format implements Parcelable {
         /* pcmEncoding= */ NO_VALUE,
         /* encoderDelay= */ NO_VALUE,
         /* encoderPadding= */ NO_VALUE,
-        language,
         /* accessibilityChannel= */ NO_VALUE,
         /* exoMediaCryptoType= */ null);
   }
@@ -688,9 +730,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         /* label= */ null,
+        /* language= */ null,
         /* selectionFlags= */ 0,
         /* roleFlags= */ 0,
-        /* bitrate= */ NO_VALUE,
+        /* averageBitrate= */ NO_VALUE,
+        /* peakBitrate= */ NO_VALUE,
         /* codecs= */ null,
         /* metadata= */ null,
         /* containerMimeType= */ null,
@@ -712,7 +756,6 @@ public final class Format implements Parcelable {
         /* pcmEncoding= */ NO_VALUE,
         /* encoderDelay= */ NO_VALUE,
         /* encoderPadding= */ NO_VALUE,
-        /* language= */ null,
         /* accessibilityChannel= */ NO_VALUE,
         /* exoMediaCryptoType= */ null);
   }
@@ -720,9 +763,11 @@ public final class Format implements Parcelable {
   /* package */ Format(
       @Nullable String id,
       @Nullable String label,
+      @Nullable String language,
       @C.SelectionFlags int selectionFlags,
       @C.RoleFlags int roleFlags,
-      int bitrate,
+      int averageBitrate,
+      int peakBitrate,
       @Nullable String codecs,
       @Nullable Metadata metadata,
       // Container specific.
@@ -748,16 +793,18 @@ public final class Format implements Parcelable {
       @C.PcmEncoding int pcmEncoding,
       int encoderDelay,
       int encoderPadding,
-      // Audio and text specific.
-      @Nullable String language,
+      // Text specific.
       int accessibilityChannel,
       // Provided by source.
       @Nullable Class<? extends ExoMediaCrypto> exoMediaCryptoType) {
     this.id = id;
     this.label = label;
+    this.language = Util.normalizeLanguageCode(language);
     this.selectionFlags = selectionFlags;
     this.roleFlags = roleFlags;
-    this.bitrate = bitrate;
+    this.averageBitrate = averageBitrate;
+    this.peakBitrate = peakBitrate;
+    this.bitrate = peakBitrate != NO_VALUE ? peakBitrate : averageBitrate;
     this.codecs = codecs;
     this.metadata = metadata;
     // Container specific.
@@ -784,8 +831,7 @@ public final class Format implements Parcelable {
     this.pcmEncoding = pcmEncoding;
     this.encoderDelay = encoderDelay == NO_VALUE ? 0 : encoderDelay;
     this.encoderPadding = encoderPadding == NO_VALUE ? 0 : encoderPadding;
-    // Audio and text specific.
-    this.language = Util.normalizeLanguageCode(language);
+    // Text specific.
     this.accessibilityChannel = accessibilityChannel;
     // Provided by source.
     this.exoMediaCryptoType = exoMediaCryptoType;
@@ -795,9 +841,12 @@ public final class Format implements Parcelable {
   /* package */ Format(Parcel in) {
     id = in.readString();
     label = in.readString();
+    language = in.readString();
     selectionFlags = in.readInt();
     roleFlags = in.readInt();
-    bitrate = in.readInt();
+    averageBitrate = in.readInt();
+    peakBitrate = in.readInt();
+    bitrate = peakBitrate != NO_VALUE ? peakBitrate : averageBitrate;
     codecs = in.readString();
     metadata = in.readParcelable(Metadata.class.getClassLoader());
     // Container specific.
@@ -828,8 +877,7 @@ public final class Format implements Parcelable {
     pcmEncoding = in.readInt();
     encoderDelay = in.readInt();
     encoderPadding = in.readInt();
-    // Audio and text specific.
-    language = in.readString();
+    // Text specific.
     accessibilityChannel = in.readInt();
     // Provided by source.
     exoMediaCryptoType = null;
@@ -839,9 +887,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        averageBitrate,
+        peakBitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -863,7 +913,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -872,9 +921,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        averageBitrate,
+        peakBitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -896,7 +947,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -905,9 +955,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        averageBitrate,
+        peakBitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -929,7 +981,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -954,9 +1005,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        /* averageBitrate= */ bitrate,
+        /* peakBitrate= */ bitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -978,7 +1031,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -1004,12 +1056,14 @@ public final class Format implements Parcelable {
     }
 
     // Prefer sample format values, but fill in from manifest if missing.
-    int bitrate = this.bitrate == NO_VALUE ? manifestFormat.bitrate : this.bitrate;
+    int averageBitrate =
+        this.averageBitrate == NO_VALUE ? manifestFormat.averageBitrate : this.averageBitrate;
+    int peakBitrate = this.peakBitrate == NO_VALUE ? manifestFormat.peakBitrate : this.peakBitrate;
     @Nullable String codecs = this.codecs;
     if (codecs == null) {
       // The manifest format may be muxed, so filter only codecs of this format's type. If we still
       // have more than one codec then we're unable to uniquely identify which codec to fill in.
-      String codecsOfType = Util.getCodecsOfType(manifestFormat.codecs, trackType);
+      @Nullable String codecsOfType = Util.getCodecsOfType(manifestFormat.codecs, trackType);
       if (Util.splitCodecs(codecsOfType).length == 1) {
         codecs = codecsOfType;
       }
@@ -1036,9 +1090,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        averageBitrate,
+        peakBitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -1060,7 +1116,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -1069,9 +1124,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        averageBitrate,
+        peakBitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -1093,7 +1150,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -1102,9 +1158,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        averageBitrate,
+        peakBitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -1126,7 +1184,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -1148,9 +1205,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        averageBitrate,
+        peakBitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -1172,7 +1231,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -1181,9 +1239,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        averageBitrate,
+        peakBitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -1205,7 +1265,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -1214,9 +1273,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        averageBitrate,
+        peakBitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -1238,7 +1299,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -1247,9 +1307,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        averageBitrate,
+        peakBitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -1271,7 +1333,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -1281,9 +1342,11 @@ public final class Format implements Parcelable {
     return new Format(
         id,
         label,
+        language,
         selectionFlags,
         roleFlags,
-        bitrate,
+        averageBitrate,
+        peakBitrate,
         codecs,
         metadata,
         containerMimeType,
@@ -1305,7 +1368,6 @@ public final class Format implements Parcelable {
         pcmEncoding,
         encoderDelay,
         encoderPadding,
-        language,
         accessibilityChannel,
         exoMediaCryptoType);
   }
@@ -1355,6 +1417,7 @@ public final class Format implements Parcelable {
       int result = 17;
       result = 31 * result + (id == null ? 0 : id.hashCode());
       result = 31 * result + (label != null ? label.hashCode() : 0);
+      result = 31 * result + (language == null ? 0 : language.hashCode());
       result = 31 * result + selectionFlags;
       result = 31 * result + roleFlags;
       result = 31 * result + bitrate;
@@ -1383,8 +1446,7 @@ public final class Format implements Parcelable {
       result = 31 * result + pcmEncoding;
       result = 31 * result + encoderDelay;
       result = 31 * result + encoderPadding;
-      // Audio and text specific.
-      result = 31 * result + (language == null ? 0 : language.hashCode());
+      // Text specific.
       result = 31 * result + accessibilityChannel;
       // Provided by source.
       result = 31 * result + (exoMediaCryptoType == null ? 0 : exoMediaCryptoType.hashCode());
@@ -1504,9 +1566,11 @@ public final class Format implements Parcelable {
   public void writeToParcel(Parcel dest, int flags) {
     dest.writeString(id);
     dest.writeString(label);
+    dest.writeString(language);
     dest.writeInt(selectionFlags);
     dest.writeInt(roleFlags);
-    dest.writeInt(bitrate);
+    dest.writeInt(averageBitrate);
+    dest.writeInt(peakBitrate);
     dest.writeString(codecs);
     dest.writeParcelable(metadata, 0);
     // Container specific.
@@ -1539,8 +1603,7 @@ public final class Format implements Parcelable {
     dest.writeInt(pcmEncoding);
     dest.writeInt(encoderDelay);
     dest.writeInt(encoderPadding);
-    // Audio and text specific.
-    dest.writeString(language);
+    // Text specific.
     dest.writeInt(accessibilityChannel);
   }