+++ library/common/src/main/java/com/google/android/exoplayer2/upstream/DataSpec.java
@@ -32,6 +32,188 @@ import java.util.Map;
  */
 public final class DataSpec {
 
+  /** Builds {@link DataSpec} instances. */
+  public static final class Builder {
+
+    @Nullable private Uri uri;
+    private long uriPositionOffset;
+    @HttpMethod private int httpMethod;
+    @Nullable private byte[] httpBody;
+    private Map<String, String> httpRequestHeaders;
+    private long position;
+    private long length;
+    @Nullable private String key;
+    @Flags private int flags;
+    @Nullable private Object customData;
+
+    /** Creates a new instance with default values. */
+    public Builder() {
+      httpMethod = HTTP_METHOD_GET;
+      httpRequestHeaders = Collections.emptyMap();
+      length = C.LENGTH_UNSET;
+    }
+
+    /**
+     * Creates a new instance to build upon the provided {@link DataSpec}.
+     *
+     * @param dataSpec The {@link DataSpec} to build upon.
+     */
+    private Builder(DataSpec dataSpec) {
+      uri = dataSpec.uri;
+      uriPositionOffset = dataSpec.uriPositionOffset;
+      httpMethod = dataSpec.httpMethod;
+      httpBody = dataSpec.httpBody;
+      httpRequestHeaders = dataSpec.httpRequestHeaders;
+      position = dataSpec.position;
+      length = dataSpec.length;
+      key = dataSpec.key;
+      flags = dataSpec.flags;
+      customData = dataSpec.customData;
+    }
+
+    /**
+     * Sets {@link DataSpec#uri}.
+     *
+     * @param uriString The {@link DataSpec#uri}.
+     * @return The builder.
+     */
+    public Builder setUri(String uriString) {
+      this.uri = Uri.parse(uriString);
+      return this;
+    }
+
+    /**
+     * Sets {@link DataSpec#uri}.
+     *
+     * @param uri The {@link DataSpec#uri}.
+     * @return The builder.
+     */
+    public Builder setUri(Uri uri) {
+      this.uri = uri;
+      return this;
+    }
+
+    /**
+     * Sets the {@link DataSpec#uriPositionOffset}. The default value is 0.
+     *
+     * @param uriPositionOffset The {@link DataSpec#uriPositionOffset}.
+     * @return The builder.
+     */
+    public Builder setUriPositionOffset(long uriPositionOffset) {
+      this.uriPositionOffset = uriPositionOffset;
+      return this;
+    }
+
+    /**
+     * Sets {@link DataSpec#httpMethod}. The default value is {@link #HTTP_METHOD_GET}.
+     *
+     * @param httpMethod The {@link DataSpec#httpMethod}.
+     * @return The builder.
+     */
+    public Builder setHttpMethod(@HttpMethod int httpMethod) {
+      this.httpMethod = httpMethod;
+      return this;
+    }
+
+    /**
+     * Sets {@link DataSpec#httpBody}. The default value is {@code null}.
+     *
+     * @param httpBody The {@link DataSpec#httpBody}.
+     * @return The builder.
+     */
+    public Builder setHttpBody(@Nullable byte[] httpBody) {
+      this.httpBody = httpBody;
+      return this;
+    }
+
+    /**
+     * Sets the {@link DataSpec#httpRequestHeaders}. The default value is an empty map.
+     *
+     * @param httpRequestHeaders The {@link DataSpec#httpRequestHeaders}.
+     * @return The builder.
+     */
+    public Builder setHttpRequestHeaders(Map<String, String> httpRequestHeaders) {
+      this.httpRequestHeaders = httpRequestHeaders;
+      return this;
+    }
+
+    /**
+     * Sets the {@link DataSpec#position}. The default value is 0.
+     *
+     * @param position The {@link DataSpec#position}.
+     * @return The builder.
+     */
+    public Builder setPosition(long position) {
+      this.position = position;
+      return this;
+    }
+
+    /**
+     * Sets the {@link DataSpec#length}. The default value is {@link C#LENGTH_UNSET}.
+     *
+     * @param length The {@link DataSpec#length}.
+     * @return The builder.
+     */
+    public Builder setLength(long length) {
+      this.length = length;
+      return this;
+    }
+
+    /**
+     * Sets the {@link DataSpec#key}. The default value is {@code null}.
+     *
+     * @param key The {@link DataSpec#key}.
+     * @return The builder.
+     */
+    public Builder setKey(@Nullable String key) {
+      this.key = key;
+      return this;
+    }
+
+    /**
+     * Sets the {@link DataSpec#flags}. The default value is 0.
+     *
+     * @param flags The {@link DataSpec#flags}.
+     * @return The builder.
+     */
+    public Builder setFlags(@Flags int flags) {
+      this.flags = flags;
+      return this;
+    }
+
+    /**
+     * Sets the {@link DataSpec#customData}. The default value is {@code null}.
+     *
+     * @param customData The {@link DataSpec#customData}.
+     * @return The builder.
+     */
+    public Builder setCustomData(@Nullable Object customData) {
+      this.customData = customData;
+      return this;
+    }
+
+    /**
+     * Builds a {@link DataSpec} with the builder's current values.
+     *
+     * @return The build {@link DataSpec}.
+     * @throws IllegalStateException If {@link #setUri} has not been called.
+     */
+    public DataSpec build() {
+      Assertions.checkStateNotNull(uri, "The uri must be set.");
+      return new DataSpec(
+          uri,
+          uriPositionOffset,
+          httpMethod,
+          httpBody,
+          httpRequestHeaders,
+          position,
+          length,
+          key,
+          flags,
+          customData);
+    }
+  }
+
   /**
    * The flags that apply to any request for data. Possible flag values are {@link
    * #FLAG_ALLOW_GZIP}, {@link #FLAG_DONT_CACHE_IF_LENGTH_UNKNOWN}, {@link
@@ -112,10 +294,26 @@ public final class DataSpec {
   public final Uri uri;
 
   /**
+   * The offset of the data located at {@link #uri} within an original resource.
+   *
+   * <p>Equal to 0 unless {@link #uri} provides access to a subset of an original resource. As an
+   * example, consider a resource that can be requested over the network and is 1000 bytes long. If
+   * {@link #uri} points to a local file that contains just bytes [200-300], then this field will be
+   * set to {@code 200}.
+   *
+   * <p>This field can be ignored except for in specific circumstances where the absolute position
+   * in the original resource is required in a {@link DataSource} chain. One example is when a
+   * {@link DataSource} needs to decrypt the content as it's read. In this case the absolute
+   * position in the original resource is typically needed to correctly initialize the decryption
+   * algorithm.
+   */
+  public final long uriPositionOffset;
+
+  /**
    * The HTTP method to use when requesting the data. This value will be ignored by non-HTTP {@link
    * DataSource} implementations.
    */
-  public final @HttpMethod int httpMethod;
+  @HttpMethod public final int httpMethod;
 
   /**
    * The HTTP request body, null otherwise. If the body is non-null, then {@code httpBody.length}
@@ -126,15 +324,16 @@ public final class DataSpec {
   /** Immutable map containing the headers to use in HTTP requests. */
   public final Map<String, String> httpRequestHeaders;
 
-  /** The absolute position of the data in the full stream. */
-  public final long absoluteStreamPosition;
-
   /**
-   * The position of the data when read from {@link #uri}.
-   * <p>
-   * Always equal to {@link #absoluteStreamPosition} unless the {@link #uri} defines the location
-   * of a subset of the underlying data.
+   * The absolute position of the data in the full stream.
+   *
+   * @deprecated Use {@link #position} except for specific use cases where the absolute position
+   *     within the original resource is required within a {@link DataSource} chain. Where the
+   *     absolute position is required, use {@code uriPositionOffset + position}.
    */
+  @Deprecated public final long absoluteStreamPosition;
+
+  /** The position of the data when read from {@link #uri}. */
   public final long position;
 
   /**
@@ -149,63 +348,103 @@ public final class DataSpec {
   @Nullable public final String key;
 
   /** Request {@link Flags flags}. */
-  public final @Flags int flags;
+  @Flags public final int flags;
 
   /**
-   * Construct a data spec for the given uri and with {@link #key} set to null.
+   * Application specific data.
+   *
+   * <p>This field is intended for advanced use cases in which applications require the ability to
+   * attach custom data to {@link DataSpec} instances. The custom data should be immutable.
+   */
+  @Nullable public final Object customData;
+
+  /**
+   * Constructs an instance.
    *
    * @param uri {@link #uri}.
    */
   public DataSpec(Uri uri) {
-    this(uri, /* flags= */ 0);
+    this(uri, /* position= */ 0, /* length= */ C.LENGTH_UNSET);
   }
 
   /**
-   * Construct a data spec for the given uri and with {@link #key} set to null.
+   * Constructs an instance.
    *
+   * @param uri {@link #uri}.
+   * @param position {@link #position}.
+   * @param length {@link #length}.
+   */
+  public DataSpec(Uri uri, long position, long length) {
+    this(
+        uri,
+        /* uriPositionOffset= */ 0,
+        HTTP_METHOD_GET,
+        /* httpBody= */ null,
+        /* httpRequestHeaders= */ Collections.emptyMap(),
+        position,
+        length,
+        /* key= */ null,
+        /* flags= */ 0,
+        /* customData= */ null);
+  }
+
+  /**
+   * Constructs an instance.
+   *
+   * @deprecated Use {@link Builder}.
    * @param uri {@link #uri}.
    * @param flags {@link #flags}.
    */
+  @SuppressWarnings("deprecation")
+  @Deprecated
   public DataSpec(Uri uri, @Flags int flags) {
     this(uri, /* position= */ 0, C.LENGTH_UNSET, /* key= */ null, flags);
   }
 
   /**
-   * Construct a data spec where {@link #position} equals {@link #absoluteStreamPosition}.
+   * Constructs an instance.
    *
+   * @deprecated Use {@link Builder}.
    * @param uri {@link #uri}.
-   * @param position {@link #position}, equal to {@link #absoluteStreamPosition}.
+   * @param position {@link #position}.
    * @param length {@link #length}.
    * @param key {@link #key}.
    */
+  @SuppressWarnings("deprecation")
+  @Deprecated
   public DataSpec(Uri uri, long position, long length, @Nullable String key) {
     this(uri, position, position, length, key, /* flags= */ 0);
   }
 
   /**
-   * Construct a data spec where {@link #position} equals {@link #absoluteStreamPosition}.
+   * Constructs an instance.
    *
+   * @deprecated Use {@link Builder}.
    * @param uri {@link #uri}.
-   * @param position {@link #position}, equal to {@link #absoluteStreamPosition}.
+   * @param position {@link #position}.
    * @param length {@link #length}.
    * @param key {@link #key}.
    * @param flags {@link #flags}.
    */
+  @SuppressWarnings("deprecation")
+  @Deprecated
   public DataSpec(Uri uri, long position, long length, @Nullable String key, @Flags int flags) {
     this(uri, position, position, length, key, flags);
   }
 
   /**
-   * Construct a data spec where {@link #position} equals {@link #absoluteStreamPosition} and has
-   * request headers.
+   * Constructs an instance.
    *
+   * @deprecated Use {@link Builder}.
    * @param uri {@link #uri}.
-   * @param position {@link #position}, equal to {@link #absoluteStreamPosition}.
+   * @param position {@link #position}, equal to {@link #position}.
    * @param length {@link #length}.
    * @param key {@link #key}.
    * @param flags {@link #flags}.
    * @param httpRequestHeaders {@link #httpRequestHeaders}
    */
+  @SuppressWarnings("deprecation")
+  @Deprecated
   public DataSpec(
       Uri uri,
       long position,
@@ -226,15 +465,18 @@ public final class DataSpec {
   }
 
   /**
-   * Construct a data spec where {@link #position} may differ from {@link #absoluteStreamPosition}.
+   * Constructs an instance where {@link #uriPositionOffset} may be non-zero.
    *
+   * @deprecated Use {@link Builder}.
    * @param uri {@link #uri}.
-   * @param absoluteStreamPosition {@link #absoluteStreamPosition}.
+   * @param absoluteStreamPosition The sum of {@link #uriPositionOffset} and {@link #position}.
    * @param position {@link #position}.
    * @param length {@link #length}.
    * @param key {@link #key}.
    * @param flags {@link #flags}.
    */
+  @SuppressWarnings("deprecation")
+  @Deprecated
   public DataSpec(
       Uri uri,
       long absoluteStreamPosition,
@@ -246,19 +488,23 @@ public final class DataSpec {
   }
 
   /**
-   * Construct a data spec by inferring the {@link #httpMethod} based on the {@code postBody}
-   * parameter. If postBody is non-null, then httpMethod is set to {@link #HTTP_METHOD_POST}. If
-   * postBody is null, then httpMethod is set to {@link #HTTP_METHOD_GET}.
+   * Construct a instance where {@link #uriPositionOffset} may be non-zero. The {@link #httpMethod}
+   * is inferred from {@code postBody}. If {@code postBody} is non-null then {@link #httpMethod} is
+   * set to {@link #HTTP_METHOD_POST}. If {@code postBody} is null then {@link #httpMethod} is set
+   * to {@link #HTTP_METHOD_GET}.
    *
+   * @deprecated Use {@link Builder}.
    * @param uri {@link #uri}.
    * @param postBody {@link #httpBody} The body of the HTTP request, which is also used to infer the
    *     {@link #httpMethod}.
-   * @param absoluteStreamPosition {@link #absoluteStreamPosition}.
+   * @param absoluteStreamPosition The sum of {@link #uriPositionOffset} and {@link #position}.
    * @param position {@link #position}.
    * @param length {@link #length}.
    * @param key {@link #key}.
    * @param flags {@link #flags}.
    */
+  @SuppressWarnings("deprecation")
+  @Deprecated
   public DataSpec(
       Uri uri,
       @Nullable byte[] postBody,
@@ -279,17 +525,20 @@ public final class DataSpec {
   }
 
   /**
-   * Construct a data spec where {@link #position} may differ from {@link #absoluteStreamPosition}.
+   * Construct a instance where {@link #uriPositionOffset} may be non-zero.
    *
+   * @deprecated Use {@link Builder}.
    * @param uri {@link #uri}.
    * @param httpMethod {@link #httpMethod}.
    * @param httpBody {@link #httpBody}.
-   * @param absoluteStreamPosition {@link #absoluteStreamPosition}.
+   * @param absoluteStreamPosition The sum of {@link #uriPositionOffset} and {@link #position}.
    * @param position {@link #position}.
    * @param length {@link #length}.
    * @param key {@link #key}.
    * @param flags {@link #flags}.
    */
+  @SuppressWarnings("deprecation")
+  @Deprecated
   public DataSpec(
       Uri uri,
       @HttpMethod int httpMethod,
@@ -312,18 +561,20 @@ public final class DataSpec {
   }
 
   /**
-   * Construct a data spec with request parameters to be used as HTTP headers inside HTTP requests.
+   * Construct a instance where {@link #uriPositionOffset} may be non-zero.
    *
+   * @deprecated Use {@link Builder}.
    * @param uri {@link #uri}.
    * @param httpMethod {@link #httpMethod}.
    * @param httpBody {@link #httpBody}.
-   * @param absoluteStreamPosition {@link #absoluteStreamPosition}.
+   * @param absoluteStreamPosition The sum of {@link #uriPositionOffset} and {@link #position}.
    * @param position {@link #position}.
    * @param length {@link #length}.
    * @param key {@link #key}.
    * @param flags {@link #flags}.
    * @param httpRequestHeaders {@link #httpRequestHeaders}.
    */
+  @Deprecated
   public DataSpec(
       Uri uri,
       @HttpMethod int httpMethod,
@@ -334,18 +585,47 @@ public final class DataSpec {
       @Nullable String key,
       @Flags int flags,
       Map<String, String> httpRequestHeaders) {
-    Assertions.checkArgument(absoluteStreamPosition >= 0);
+    this(
+        uri,
+        /* uriPositionOffset= */ absoluteStreamPosition - position,
+        httpMethod,
+        httpBody,
+        httpRequestHeaders,
+        position,
+        length,
+        key,
+        flags,
+        /* customData= */ null);
+  }
+
+  @SuppressWarnings("deprecation")
+  private DataSpec(
+      Uri uri,
+      long uriPositionOffset,
+      @HttpMethod int httpMethod,
+      @Nullable byte[] httpBody,
+      Map<String, String> httpRequestHeaders,
+      long position,
+      long length,
+      @Nullable String key,
+      @Flags int flags,
+      @Nullable Object customData) {
+    // TODO: Replace this assertion with a stricter one checking "uriPositionOffset >= 0", after
+    // validating there are no violations in ExoPlayer and 1P apps.
+    Assertions.checkArgument(uriPositionOffset + position >= 0);
     Assertions.checkArgument(position >= 0);
     Assertions.checkArgument(length > 0 || length == C.LENGTH_UNSET);
     this.uri = uri;
+    this.uriPositionOffset = uriPositionOffset;
     this.httpMethod = httpMethod;
-    this.httpBody = (httpBody != null && httpBody.length != 0) ? httpBody : null;
-    this.absoluteStreamPosition = absoluteStreamPosition;
+    this.httpBody = httpBody != null && httpBody.length != 0 ? httpBody : null;
+    this.httpRequestHeaders = Collections.unmodifiableMap(new HashMap<>(httpRequestHeaders));
     this.position = position;
+    this.absoluteStreamPosition = uriPositionOffset + position;
     this.length = length;
     this.key = key;
     this.flags = flags;
-    this.httpRequestHeaders = Collections.unmodifiableMap(new HashMap<>(httpRequestHeaders));
+    this.customData = customData;
   }
 
   /**
@@ -363,6 +643,11 @@ public final class DataSpec {
    */
   public final String getHttpMethodString() {
     return getStringForHttpMethod(httpMethod);
+  }
+
+  /** Returns a {@link DataSpec.Builder} initialized with the values of this instance. */
+  public DataSpec.Builder buildUpon() {
+    return new Builder(this);
   }
 
   /**
@@ -389,14 +674,15 @@ public final class DataSpec {
     } else {
       return new DataSpec(
           uri,
+          uriPositionOffset,
           httpMethod,
           httpBody,
-          absoluteStreamPosition + offset,
+          httpRequestHeaders,
           position + offset,
           length,
           key,
           flags,
-          httpRequestHeaders);
+          customData);
     }
   }
 
@@ -409,14 +695,15 @@ public final class DataSpec {
   public DataSpec withUri(Uri uri) {
     return new DataSpec(
         uri,
+        uriPositionOffset,
         httpMethod,
         httpBody,
-        absoluteStreamPosition,
+        httpRequestHeaders,
         position,
         length,
         key,
         flags,
-        httpRequestHeaders);
+        customData);
   }
 
   /**
@@ -429,14 +716,15 @@ public final class DataSpec {
   public DataSpec withRequestHeaders(Map<String, String> httpRequestHeaders) {
     return new DataSpec(
         uri,
+        uriPositionOffset,
         httpMethod,
         httpBody,
-        absoluteStreamPosition,
+        httpRequestHeaders,
         position,
         length,
         key,
         flags,
-        httpRequestHeaders);
+        customData);
   }
 
   /**
@@ -452,14 +740,15 @@ public final class DataSpec {
     httpRequestHeaders.putAll(additionalHttpRequestHeaders);
     return new DataSpec(
         uri,
+        uriPositionOffset,
         httpMethod,
         httpBody,
-        absoluteStreamPosition,
+        httpRequestHeaders,
         position,
         length,
         key,
         flags,
-        httpRequestHeaders);
+        customData);
   }
 
   @Override