+++ library/core/src/main/java/com/google/android/exoplayer2/audio/MediaCodecAudioRenderer.java
@@ -22,7 +22,6 @@ import android.media.MediaCrypto;
 import android.media.MediaFormat;
 import android.media.audiofx.Virtualizer;
 import android.os.Handler;
-import androidx.annotation.CallSuper;
 import androidx.annotation.Nullable;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.ExoPlaybackException;
@@ -43,7 +42,6 @@ import com.google.android.exoplayer2.mediacodec.MediaCodecUtil;
 import com.google.android.exoplayer2.mediacodec.MediaCodecUtil.DecoderQueryException;
 import com.google.android.exoplayer2.mediacodec.MediaFormatUtil;
 import com.google.android.exoplayer2.source.MediaSource;
-import com.google.android.exoplayer2.util.Log;
 import com.google.android.exoplayer2.util.MediaClock;
 import com.google.android.exoplayer2.util.MimeTypes;
 import com.google.android.exoplayer2.util.Util;
@@ -71,13 +69,6 @@ import java.util.List;
  */
 public class MediaCodecAudioRenderer extends MediaCodecRenderer implements MediaClock {
 
-  /**
-   * Maximum number of tracked pending stream change times. Generally there is zero or one pending
-   * stream change. We track more to allow for pending changes that have fewer samples than the
-   * codec latency.
-   */
-  private static final int MAX_PENDING_STREAM_CHANGE_COUNT = 10;
-
   private static final String TAG = "MediaCodecAudioRenderer";
   /**
    * Custom key used to indicate bits per sample by some decoders on Vivo devices. For example
@@ -88,7 +79,6 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
   private final Context context;
   private final EventDispatcher eventDispatcher;
   private final AudioSink audioSink;
-  private final long[] pendingStreamChangeTimesUs;
 
   private int codecMaxInputSize;
   private boolean passthroughEnabled;
@@ -99,8 +89,8 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
   private long currentPositionUs;
   private boolean allowFirstBufferPositionDiscontinuity;
   private boolean allowPositionDiscontinuity;
-  private long lastInputTimeUs;
-  private int pendingStreamChangeCount;
+
+  @C.Encoding private int audioSinkEncoding;
 
   /**
    * @param context A context.
@@ -354,8 +344,6 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
         /* assumedMinimumCodecOperatingRate= */ 44100);
     this.context = context.getApplicationContext();
     this.audioSink = audioSink;
-    lastInputTimeUs = C.TIME_UNSET;
-    pendingStreamChangeTimesUs = new long[MAX_PENDING_STREAM_CHANGE_COUNT];
     eventDispatcher = new EventDispatcher(eventHandler, eventListener);
     audioSink.setListener(new AudioSinkListener());
   }
@@ -602,6 +590,7 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
       // TODO(internal: b/145658993) Use outputFormat instead.
       throw createRendererException(e, inputFormat);
     }
+    audioSinkEncoding = encoding;
   }
 
   /**
@@ -667,30 +656,12 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
   }
 
   @Override
-  protected void onStreamChanged(Format[] formats, long offsetUs) throws ExoPlaybackException {
-    super.onStreamChanged(formats, offsetUs);
-    if (lastInputTimeUs != C.TIME_UNSET) {
-      if (pendingStreamChangeCount == pendingStreamChangeTimesUs.length) {
-        Log.w(
-            TAG,
-            "Too many stream changes, so dropping change at "
-                + pendingStreamChangeTimesUs[pendingStreamChangeCount - 1]);
-      } else {
-        pendingStreamChangeCount++;
-      }
-      pendingStreamChangeTimesUs[pendingStreamChangeCount - 1] = lastInputTimeUs;
-    }
-  }
-
-  @Override
   protected void onPositionReset(long positionUs, boolean joining) throws ExoPlaybackException {
     super.onPositionReset(positionUs, joining);
     audioSink.flush();
     currentPositionUs = positionUs;
     allowFirstBufferPositionDiscontinuity = true;
     allowPositionDiscontinuity = true;
-    lastInputTimeUs = C.TIME_UNSET;
-    pendingStreamChangeCount = 0;
   }
 
   @Override
@@ -709,8 +680,6 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
   @Override
   protected void onDisabled() {
     try {
-      lastInputTimeUs = C.TIME_UNSET;
-      pendingStreamChangeCount = 0;
       audioSink.flush();
     } finally {
       try {
@@ -769,22 +738,12 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
       }
       allowFirstBufferPositionDiscontinuity = false;
     }
-    lastInputTimeUs = Math.max(buffer.timeUs, lastInputTimeUs);
   }
 
-  @CallSuper
   @Override
-  protected void onProcessedOutputBuffer(long presentationTimeUs) {
-    while (pendingStreamChangeCount != 0 && presentationTimeUs >= pendingStreamChangeTimesUs[0]) {
-      audioSink.handleDiscontinuity();
-      pendingStreamChangeCount--;
-      System.arraycopy(
-          pendingStreamChangeTimesUs,
-          /* srcPos= */ 1,
-          pendingStreamChangeTimesUs,
-          /* destPos= */ 0,
-          pendingStreamChangeCount);
-    }
+  protected void onProcessedStreamChange() {
+    super.onProcessedStreamChange();
+    audioSink.handleDiscontinuity();
   }
 
   @Override
@@ -795,6 +754,7 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
       ByteBuffer buffer,
       int bufferIndex,
       int bufferFlags,
+      int sampleCount,
       long bufferPresentationTimeUs,
       boolean isDecodeOnlyBuffer,
       boolean isLastBuffer,
@@ -803,8 +763,8 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
     if (codecNeedsEosBufferTimestampWorkaround
         && bufferPresentationTimeUs == 0
         && (bufferFlags & MediaCodec.BUFFER_FLAG_END_OF_STREAM) != 0
-        && lastInputTimeUs != C.TIME_UNSET) {
-      bufferPresentationTimeUs = lastInputTimeUs;
+        && getLargestQueuedPresentationTimeUs() != C.TIME_UNSET) {
+      bufferPresentationTimeUs = getLargestQueuedPresentationTimeUs();
     }
 
     if (passthroughEnabled && (bufferFlags & MediaCodec.BUFFER_FLAG_CODEC_CONFIG) != 0) {
@@ -820,16 +780,25 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
       return true;
     }
 
+    boolean fullyConsumed;
     try {
-      if (audioSink.handleBuffer(buffer, bufferPresentationTimeUs)) {
-        codec.releaseOutputBuffer(bufferIndex, false);
-        decoderCounters.renderedOutputBufferCount++;
-        return true;
+      if (Util.isEncodingLinearPcm(audioSinkEncoding)) {
+        fullyConsumed = audioSink.handleBuffer(buffer, bufferPresentationTimeUs);
+      } else {
+        fullyConsumed =
+            audioSink.handleEncodedBuffer(buffer, bufferPresentationTimeUs, sampleCount);
       }
     } catch (AudioSink.InitializationException | AudioSink.WriteException e) {
       // TODO(internal: b/145658993) Use outputFormat instead.
       throw createRendererException(e, inputFormat);
     }
+
+    if (fullyConsumed) {
+      codec.releaseOutputBuffer(bufferIndex, false);
+      decoderCounters.renderedOutputBufferCount++;
+      return true;
+    }
+
     return false;
   }