+++ RELEASENOTES.md
@@ -3,14 +3,18 @@
 ### dev-v2 (not yet released) ###
 
 * Core library:
+  * Add fields `videoFrameProcessingOffsetUsSum` and
+    `videoFrameProcessingOffsetUsCount` in `DecoderCounters` to compute
+    the average video frame processing offset.
   * Add playlist API ([#6161](https://github.com/google/ExoPlayer/issues/6161)).
   * Add `play` and `pause` methods to `Player`.
   * Add `Player.getCurrentLiveOffset` to conveniently return the live offset.
   * Add `Player.onPlayWhenReadyChanged` with reasons.
   * Add `Player.onPlaybackStateChanged` and deprecate
     `Player.onPlayerStateChanged`.
-  * Deprecate and rename getPlaybackError to getPlayerError for consistency.
-  * Deprecate and rename onLoadingChanged to onIsLoadingChanged for consistency.
+  * Deprecate and rename `getPlaybackError` to `getPlayerError` for consistency.
+  * Deprecate and rename `onLoadingChanged` to `onIsLoadingChanged` for
+    consistency.
   * Make `MediaSourceEventListener.LoadEventInfo` and
     `MediaSourceEventListener.MediaLoadData` top-level classes.
   * Rename `MediaCodecRenderer.onOutputFormatChanged` to
@@ -19,6 +23,13 @@
   * Move player message-related constants from `C` to `Renderer`, to avoid
     having the constants class depend on player/renderer classes.
   * Split out `common` and `extractor` submodules.
+  * Allow to explicitly send `PlayerMessage`s at the end of a stream.
+  * Add `DataSpec.Builder` and deprecate most `DataSpec` constructors.
+  * Add `DataSpec.customData` to allow applications to pass custom data through
+    `DataSource` chains.
+  * Move the handling of encoded buffers in audio passthrough from
+    `AudioSink.handleBuffer` to `AudioSink.handleEncodedBuffer` to allow
+    passing multiple encoded frames in one buffer.
 * Text:
   * Parse `<ruby>` and `<rt>` tags in WebVTT subtitles (rendering is coming
     later).